refactor(remote-config): migrate rctelemetryreporter component to v2
