+++ comp/remote-config/rctelemetryreporter/def/component.go
